fix/birthday-widget: Polaroid no longer rotates.
An issue was raised for the birthday widget polaroid not suiting the current \'square\' style theme of the dashboard.
This has caused some distress in the people viewing the polaroid which did not align with every other widget.

This has been fixed by removing the `style={{ transform: \'rotate(...)\'}}` line from the `Polaroid.jsx`.
By removing this, there will still be rotation of different birthdays, however with no actual rotation transformation of the polaroid.